Refactor Response writing process inspiring from standard library
Introduced the ResponseWriter interface instead of letting the author
of handlers directly modifying the response. It made the default cases
really easy, considering the hanlder won't even have to set a status or
any header. Moreoever, now if the handler wants, it can stream the content
instead of loading the entire body into the Response object, as long as it is
setting `Content-Length` header appropriately.

Note another interesting fact! In order to make the normal case simple,
we let the handler even skip the `Content-Type` header for `text/plain` header.
We also let them skip `Content-Length` if all they want to send is just a one chunk of data.
We can do that because we can provide the length ourselves from that chunk's length.