Allow handler registration on the sever struct
The registration syntax is:
server.Register("<HTTP_VERB> <Path>", handlerFunc)

The path can optionally content a path variable definition in the form of {name}.
Although currently the request object does not parse the path variable itself.

The pattern matching rule is fairly simple at this moment. It takes the registered
handlers one by one as it is registered on the object and tries to match the request
path with that handler's pattern. The first one that matches the path is considered to
be the handler.